+++ packages/rekapager-core/src/Batch/Event/AfterProcessEvent.php
@@ -15,41 +15,7 @@ namespace Rekalogika\Rekapager\Batch\Event;
 
 final class AfterProcessEvent
 {
-    /**
-     * @param int<0,max> $itemsProcessed
-     * @param int<0,max> $pagesProcessed
-     */
-    public function __construct(
-        private readonly float $processEndTime,
-        private readonly float $processDuration,
-        private readonly int $itemsProcessed,
-        private readonly int $pagesProcessed,
-    ) {
-    }
-
-    /**
-     * @return int<0,max>
-     */
-    public function getItemsProcessed(): int
+    public function __construct()
     {
-        return $this->itemsProcessed;
-    }
-
-    /**
-     * @return int<0,max>
-     */
-    public function getPagesProcessed(): int
-    {
-        return $this->pagesProcessed;
-    }
-
-    public function getProcessEndTime(): float
-    {
-        return $this->processEndTime;
-    }
-
-    public function getProcessDuration(): float
-    {
-        return $this->processDuration;
     }
 }