+++ packages/rekapager-core/src/Batch/Implementation/DefaultBatchProcessFactory.php
@@ -13,8 +13,6 @@ declare(strict_types=1);
 
 namespace Rekalogika\Rekapager\Batch\Implementation;
 
-use Psr\Log\LoggerInterface;
-use Psr\Log\NullLogger;
 use Rekalogika\Contracts\Rekapager\PageableInterface;
 use Rekalogika\Rekapager\Batch\BatchProcess;
 use Rekalogika\Rekapager\Batch\BatchProcessFactoryInterface;
@@ -23,13 +21,9 @@ use Rekalogika\Rekapager\Contracts\PageIdentifierEncoderResolverInterface;
 
 final class DefaultBatchProcessFactory implements BatchProcessFactoryInterface
 {
-    private readonly LoggerInterface $logger;
-
     public function __construct(
         private readonly PageIdentifierEncoderResolverInterface $pageableIdentifierResolver,
-        ?LoggerInterface $logger = null
     ) {
-        $this->logger = $logger ?? new NullLogger();
     }
 
     public function createBatchProcess(
@@ -40,7 +34,6 @@ final class DefaultBatchProcessFactory implements BatchProcessFactoryInterface
             pageable: $pageable,
             batchProcessor: $batchProcessor,
             pageableIdentifierResolver: $this->pageableIdentifierResolver,
-            logger: $this->logger
         );
     }
 }