+++ packages/rekapager-core/src/Batch/BatchProcessorInterface.php
@@ -19,6 +19,7 @@ use Rekalogika\Rekapager\Batch\Event\BeforePageEvent;
 use Rekalogika\Rekapager\Batch\Event\BeforeProcessEvent;
 use Rekalogika\Rekapager\Batch\Event\InterruptEvent;
 use Rekalogika\Rekapager\Batch\Event\ItemEvent;
+use Rekalogika\Rekapager\Batch\Event\TimeLimitEvent;
 
 /**
  * @template TKey of array-key
@@ -60,4 +61,9 @@ interface BatchProcessorInterface
      * @param InterruptEvent<TKey,T> $event
      */
     public function onInterrupt(InterruptEvent $event): void;
+
+    /**
+     * @param TimeLimitEvent<TKey,T> $event
+     */
+    public function onTimeLimit(TimeLimitEvent $event): void;
 }